refactor(order): validator의 구현체를 table 패키지로 두어 order -> table 참조 끊도록 수정
