+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -14,6 +14,7 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderMenu;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.support.IntegrationTest;
+import kitchenpos.table.application.OrderValidatorImpl;
 import kitchenpos.table.domain.OrderTable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -27,7 +28,7 @@ class OrderServiceTest extends IntegrationTest {
     private OrderService orderService;
 
     @Autowired
-    private OrderValidator orderValidator;
+    private OrderValidatorImpl orderValidator;
 
     private OrderTable nonEmptyOrderTable;
     private OrderLineItem orderLineItem;
